exercicio 1

+++ 0x13/resumo_pedido/Pedido.java
@@ -1,6 +1,6 @@
 public class Pedido {
     private int codigo;
-    private int peso;
+    private double peso;
     private double total;
     private Frete frete;
 
@@ -18,11 +18,11 @@ public class Pedido {
         this.codigo = codigo;
     }
 
-    public int getPeso() {
+    public double getPeso() {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         this.peso = peso;
     }
